+++ sdk/go/client/client.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/coretexos/cap/sdk/go"
-	agentv1 "github.com/coretexos/cap/cortex/agent/v1"
+	agentv1 "github.com/coretexos/cap/go/cortex/agent/v1"
 	"github.com/nats-io/nats.go"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/timestamppb"